+++ checks/normandy/recipe_signatures.py
